+++ .idea/misc.xml
@@ -3,16 +3,13 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
-        <entry key="..\:/Proyecto_Final/app/src/main/res/layout/activity_crear_factura.xml" value="0.21965579710144928" />
-        <entry key="..\:/Proyecto_Final/app/src/main/res/layout/activity_main.xml" value="0.3555253623188406" />
-        <entry key="..\:/Proyecto_Final/app/src/main/res/layout/activity_registro.xml" value="0.3555253623188406" />
-        <entry key="..\:/Proyecto_Final/app/src/main/res/layout/item_layout_emisor.xml" value="0.3555253623188406" />
-        <entry key="..\:/Proyecto_Final/app/src/main/res/layout/item_layout_receptor.xml" value="0.3555253623188406" />
-        <entry key="..\:/UIA/Labs/Proyecto_Final/app/src/main/res/layout/activity_crear_factura.xml" value="0.16847826086956522" />
+        <entry key="..\:/UIA/Labs/Proyecto_Final/app/src/main/res/layout/activity_crear_factura.xml" value="0.25" />
         <entry key="..\:/UIA/Labs/Proyecto_Final/app/src/main/res/layout/activity_inicio.xml" value="0.25" />
         <entry key="..\:/UIA/Labs/Proyecto_Final/app/src/main/res/layout/activity_main.xml" value="0.16847826086956522" />
         <entry key="..\:/UIA/Labs/Proyecto_Final/app/src/main/res/layout/activity_registro.xml" value="0.16847826086956522" />
         <entry key="..\:/UIA/Labs/Proyecto_Final/app/src/main/res/layout/activity_termcond.xml" value="0.16847826086956522" />
+        <entry key="..\:/UIA/Labs/Proyecto_Final/app/src/main/res/layout/item_layout_emisor.xml" value="0.25" />
+        <entry key="..\:/UIA/Labs/Proyecto_Final/app/src/main/res/layout/item_layout_lista.xml" value="0.20416666666666666" />
         <entry key="..\:/Users/DELL/Desktop/Lenguajes/Proyecto/Proyecto_Final/app/src/main/res/layout/activity_inicio.xml" value="0.17074275362318841" />
         <entry key="..\:/Users/DELL/Desktop/Lenguajes/Proyecto/Proyecto_Final/app/src/main/res/layout/activity_main.xml" value="0.17074275362318841" />
         <entry key="..\:/Users/DELL/Desktop/Lenguajes/Proyecto/Proyecto_Final/app/src/main/res/layout/activity_registro.xml" value="0.17074275362318841" />
