+++ app/src/main/java/com/example/proyecto_final/adapter_emisor.kt
@@ -6,28 +6,27 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import android.widget.ImageView
-import android.widget.Spinner
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.squareup.picasso.Picasso
 
 class adapter_emisor : RecyclerView.Adapter<adapter_emisor.ViewHolder>()  {
-    var emisores:MutableList<RegistroFacturaActivity.PERSONA> = ArrayList()
+    var emisores:MutableList<CapaDatos.PERSONA> = ArrayList()
 
     lateinit var context: Context
 
-    fun adapter_emisor(emisores: MutableList<RegistroFacturaActivity.PERSONA>, context: Context){
+    fun adapter_emisor(emisores: MutableList<CapaDatos.PERSONA>, context: Context){
         this.emisores = emisores
         this.context = context
 
     }
-    override fun onBindViewHolder(holder: adapter_emisor.ViewHolder, position: Int) {
+    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         val item = emisores.get(position)
         holder.bind(item, context)
     }
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): adapter_emisor.ViewHolder {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val layoutInflater = LayoutInflater.from(parent.context)
-        return adapter_emisor.ViewHolder(layoutInflater.inflate(R.layout.item_layout_emisor, parent, false))
+        return ViewHolder(layoutInflater.inflate(R.layout.item_layout_emisor, parent, false))
     }
 
     override fun getItemCount(): Int {
@@ -42,7 +41,7 @@ class adapter_emisor : RecyclerView.Adapter<adapter_emisor.ViewHolder>()  {
         val botonselecciona = view.findViewById(R.id.btnselecciona) as Button
 
 
-        fun bind(emisor: RegistroFacturaActivity.PERSONA, context: Context){
+        fun bind(emisor: CapaDatos.PERSONA, context: Context){
             nombre.text = emisor.NOMBRE1
             numero.text = emisor.NUMERO1
             correo.text= emisor.CORREOELECTRONICO1
@@ -53,8 +52,4 @@ class adapter_emisor : RecyclerView.Adapter<adapter_emisor.ViewHolder>()  {
             Picasso.with(context).load(url).into(this)
         }
     }
-
-
-
-
 }
