+++ app/src/main/res/layout/item_layout_emisor.xml
@@ -10,10 +10,10 @@
         android:id="@+id/numero"
         android:layout_width="100dp"
         android:layout_height="100dp"
-        android:layout_marginEnd="10dp"
         android:layout_marginRight="10dp"
         android:scaleType="centerCrop"
         android:textSize="40sp" />
+
     <TextView
         android:id="@+id/nombre"
         android:layout_width="wrap_content"
@@ -38,8 +38,8 @@
         android:id="@+id/btnselecciona"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Seleccionar este emisor"
-        android:layout_alignParentBottom="true"
+        android:layout_alignParentEnd="true"
         android:layout_alignParentRight="true"
-        android:layout_alignParentEnd="true"/>
+        android:layout_alignParentBottom="true"
+        android:text="Seleccionar este emisor" />
 </RelativeLayout>
